Lab4GPS/Lab4GPS/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.2
whitenoise==6.8.2
